pyproject.toml
[project]
name = "book-gen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.27",
    "langchain-deepseek>=0.1.4",
    "langchain-openai>=0.3.33",
    "langgraph>=0.6.7",
    "openai>=1.106.1",
    "python-dotenv>=1.1.1",
    "xhtml2pdf>=0.2.17",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "mypy>=1.17.1",
    "ruff>=0.13.0",
]
